# Proof-of-concept: speeding up gemm reference kernel