@bartoldeman 2025-03-10
@devinamatthews I created a `zen4` config using this reference kernel but it does mean adding it in various places through the general source code as well, not simply adding a few files under `config/zen4`. That said, `bli_kernel_defs_<arch>.h` could be a place to define `CACHELINE_SIZE` and `TAIL_NITER` perhaps. Right now I can already pass MR and NR to configure but would it be a good idea to be able to do that for some other constants too, to easily create a tuned generic kernel, without going more "heavy duty"? This is why it's POC.. I really am not familiar enough how to do that yet.